pyproject.toml
[project]
name = "llm-code-deployer1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "openai>=2.3.0",
    "pydantic>=2.12.2",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]
